+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 # KdbInsideBrains Changelog
 
-## [0.20.1]
+## [0.21.0]
 
-### Changed
+### Added
 
-- Charting improvements and context menu (for measure tool) added
+- Dark icons set added with IntelliJ Idea colors
+- Charting is ready now:
+  - New Magnet/Snap tool has been added with ability to snap to line, vertex or disable the tool (default)
+  - Values Tool shows only series values and no range values anymore
+  - Measure Tool has context menu with remove/clear ability
 
 ## [0.20.0]